% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_biopath_score}
\alias{get_biopath_score}
\title{Get biopath score by calculate the gene expression data}
\usage{
get_biopath_score(data, biopath_genes, normals = NULL, ranks = NULL,
  attempts = 50, maximize_stability = F, min_exp = 2, min_std = 0.1)
}
\arguments{
\item{data}{a matrix, row represent gene, col represent sample, must contain row names(Gene ID) and col names(Sample ID)}

\item{biopath_genes}{a vector, the Gene ID of a bio-pathway you want to caculate.}

\item{normals}{a vector, TURE or FALSE of the samples, if sample is Normal then TRUE, else is Tumor then FALSE}

\item{ranks}{a vector or NULL, if not have normals, rank your samples}

\item{attempts}{a non-negative integer, default 50, iter times}

\item{maximize_stability}{maximize_stability=T, then algorithm will run more times to maximize score stability, very slow.}

\item{min_exp}{use min_exp replace gene expression which lower than it, if min_exp = NULL, than not replace, default 2}

\item{min_std}{use min std to cut-off gene whose std below than it, min_std must > 0, default 0.1}
}
\description{
Get biopath score by calculate the gene expression data
}
\examples{
library(biopathscore)
data("dat_brca")
data('v_brca_normals')
data('l_kegg_gs_min')
test_pathway = l_kegg_gs_min$KEGG_DNA_REPLICATION
pds = get_biopath_score(data = dat_brca, biopath_genes = test_pathway, normals = v_brca_normals, maximize_stability = F)
pds2 = get_biopath_score(data = dat_brca, biopath_genes = test_pathway, normals = v_brca_normals, maximize_stability = T)

plot3D_lpc(pds, v_brca_normals,drawLine = F, outputHTML = F )
plot3D_lpc(pds2, v_brca_normals,drawLine = F, outputHTML = F )
}
