% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpc.output.R
\name{plot3D.lpc}
\alias{plot3D.lpc}
\title{Plot samples LPC line and scatter}
\usage{
plot3D.lpc(res.lpc, normals, outputHTML = T, outputDir = "./plot",
  fileName = "plot")
}
\arguments{
\item{res.lpc}{data after run get_biopath_score}

\item{normals}{a vector of the samples, if samples is Normal then TRUE, else Tumor then FALSE}

\item{outputHTML}{if need output this plot as a html file is TRUE, if not then FALSE}

\item{outputDir}{if outputHTML=T, this is availiable, the output file directory}

\item{fileName}{if outputHTML=T, this is availiable, the output file name.}
}
\value{
not have a return
}
\description{
Plot samples LPC line and scatter
}
