% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpc.output.R
\name{plot3D.lpc2}
\alias{plot3D.lpc2}
\title{Plot samples LPC line and scatter}
\usage{
plot3D.lpc2(res.lpc, normals, outputHTML = T, outputDir = "./plot",
  fileName = "plot")
}
\arguments{
\item{res.lpc}{data after run get_biopath_score}

\item{normals}{a vector of the samples, if samples is Normal then TRUE, else Tumor then FALSE}
}
\value{
not have a return
}
\description{
Plot samples LPC line and scatter
}
\examples{

}
