% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpc.output.R
\name{plot3D_lpc}
\alias{plot3D_lpc}
\title{Plot samples LPC line and scatter}
\usage{
plot3D_lpc(res.lpc, normals, drawLine = F, outputHTML = F,
  outputDir = "./plot", fileName = "plot")
}
\arguments{
\item{res.lpc}{data after run get_biopath_score}

\item{normals}{a vector of the samples, if samples is Normal then TRUE, else Tumor then FALSE}

\item{outputHTML}{if need output this plot as a html file is TRUE, if not then FALSE}

\item{outputDir}{if outputHTML=T, this is availiable, the output file directory}

\item{fileName}{if outputHTML=T, this is availiable, the output file name.}
}
\value{
not have a return
}
\description{
Plot samples LPC line and scatter
}
\examples{
library(biopathscore)
data("dat_brca")
data('v_brca_normals')
data('l_kegg_gs_min')
test_pathway = l_kegg_gs_min$KEGG_DNA_REPLICATION
pds = get_biopath_score(data = dat_brca, biopath_genes = test_pathway, normals = v_brca_normals, maximize_stability = F)
pds2 = get_biopath_score(data = dat_brca, biopath_genes = test_pathway, normals = v_brca_normals, maximize_stability = T)

plot3D_lpc(pds, v_brca_normals,drawLine = F, outputHTML = F )
plot3D_lpc(pds2, v_brca_normals,drawLine = F, outputHTML = F )

}
